Retrieves diagnoses containing a description.

Returns a dataframe of ICD code and descriptions for all entries including any supplied keyword.

ukb_icd_keyword(description, icd.version = 10)

Arguments

description
A character vector of one or more keywords to be looked up in the ICD descriptions, e.g., "cardio", c("cardio", "lymphoma"). Each keyword can be a regular expression, e.g. "lymph*".

icd.version
The ICD version (or revision) number, 9 or 10. Default = 10.

See also

Examples

ukb_icd_keyword("cardio", icd.version = 10)

#>                             code
#> 1                           B334
#> 2                           A439
#> 3                           A438
#> 4                           A431
#> 5                           A430
#> 6                            A43
#> 7  Block P20-P29(not selectable)
#> 8                           I250
#> 9                           I255
#> 10                          I420
#> 11                          I421
#> 12                          I422
#> 13                          I425
#> 14                          I426
#> 15                          I428
#> 16                          I970
#> 17                          P298
#> 18                          R943
#> 19                           T46
#> 20                          T469
#> 21                           Y52
#> 22                          Y529
#> 23                          Z035
#> 24                          Z136
#>                                                                              meaning
#> 1                          B33.4 Hantavirus (cardio-)pulmonary syndrome [HPS] [HCPS]
#> 2                                                     A43.9 Nocardiosis, unspecified
#> 3                                                   A43.8 Other forms of nocardiosis
#> 4                                                        A43.1 Cutaneous nocardiosis
#> 5                                                        A43.0 Pulmonary nocardiosis
#> 6                                                                    A43 Nocardiosis
#> 7  P20-P29 Respiratory and cardiovascular disorders specific to the perinatal period
#> 8                         I25.0 Atherosclerotic cardiovascular disease, so described
#> 9                                                     I25.5 Ischaemic cardiomyopathy
#> 10                                                      I42.0 Dilated cardiomyopathy
#> 11                                     I42.1 Obstructive hypertrophic cardiomyopathy
#> 12                                           I42.2 Other hypertrophic cardiomyopathy
#> 13                                            I42.5 Other restrictive cardiomyopathy
#> 14                                                    I42.6 Alcoholic cardiomyopathy
#> 15                                                      I42.8 Other cardiomyopathies
#> 16                                                     I97.0 Postcardiotomy syndrome
#> 17          P29.8 Other cardiovascular disorders originating in the perinatal period
#> 18                         R94.3 Abnormal results of cardiovascular function studies
#> 19             T46 Poisoning by agents primarily affecting the cardiovascular system
#> 20  T46.9 Other and unspecified agents primarily affecting the cardiovascular system
#> 21                          Y52 Agents primarily affecting the cardiovascular system
#> 22  Y52.9 Other and unspecified agents primarily affecting the cardiovascular system
#> 23                     Z03.5 Observation for other suspected cardiovascular diseases
#> 24                  Z13.6 Special screening examination for cardiovascular disorders