Retrieves diagnoses containing a description.

Returns a dataframe of ICD code and descriptions for all entries including any supplied keyword.

ukb_icd_keyword(description, icd.version = 10, ignore.case = TRUE)

Arguments

description

A character vector of one or more keywords to be looked up in the ICD descriptions, e.g., "cardio", c("cardio", "lymphoma"). Each keyword can be a regular expression, e.g. "lymph*".

icd.version

The ICD version (or revision) number, 9 or 10. Default = 10.

ignore.case

If `TRUE` (default), case is ignored during matching; if `FALSE`, the matching is case sensitive.

See also

Examples

ukb_icd_keyword("cardio", icd.version = 10)
#>                             code
#> 1                           B334
#> 2                           A520
#> 3                           A439
#> 4                           A438
#> 5                           A431
#> 6                           A430
#> 7                            A43
#> 8  Block P20-P29(not selectable)
#> 9                           I250
#> 10                          I255
#> 11                           I42
#> 12                          I420
#> 13                          I421
#> 14                          I422
#> 15                          I425
#> 16                          I426
#> 17                          I427
#> 18                          I428
#> 19                          I429
#> 20                           I43
#> 21                          I430
#> 22                          I431
#> 23                          I432
#> 24                          I438
#> 25                          I516
#> 26                          I517
#> 27                          I970
#> 28                          I980
#> 29                          I981
#> 30                          O903
#> 31                           P29
#> 32                          P298
#> 33                          P299
#> 34                          R570
#> 35                          R943
#> 36                           T46
#> 37                          T469
#> 38                           Y52
#> 39                          Y529
#> 40                           Y71
#> 41                          Z035
#> 42                          Z136
#>                                                                                           meaning
#> 1                                       B33.4 Hantavirus (cardio-)pulmonary syndrome [HPS] [HCPS]
#> 2                                                                   A52.0 Cardiovascular syphilis
#> 3                                                                  A43.9 Nocardiosis, unspecified
#> 4                                                                A43.8 Other forms of nocardiosis
#> 5                                                                     A43.1 Cutaneous nocardiosis
#> 6                                                                     A43.0 Pulmonary nocardiosis
#> 7                                                                                 A43 Nocardiosis
#> 8               P20-P29 Respiratory and cardiovascular disorders specific to the perinatal period
#> 9                                      I25.0 Atherosclerotic cardiovascular disease, so described
#> 10                                                                 I25.5 Ischaemic cardiomyopathy
#> 11                                                                             I42 Cardiomyopathy
#> 12                                                                   I42.0 Dilated cardiomyopathy
#> 13                                                  I42.1 Obstructive hypertrophic cardiomyopathy
#> 14                                                        I42.2 Other hypertrophic cardiomyopathy
#> 15                                                         I42.5 Other restrictive cardiomyopathy
#> 16                                                                 I42.6 Alcoholic cardiomyopathy
#> 17                                    I42.7 Cardiomyopathy due to drugs and other external agents
#> 18                                                                   I42.8 Other cardiomyopathies
#> 19                                                              I42.9 Cardiomyopathy, unspecified
#> 20                                            I43 Cardiomyopathy in diseases classified elsewhere
#> 21                 I43.0 Cardiomyopathy in infectious and parasitic diseases classified elsewhere
#> 22                                                     I43.1 Cardiomyopathy in metabolic diseases
#> 23                                                   I43.2 Cardiomyopathy in nutritional diseases
#> 24                                    I43.8 Cardiomyopathy in other diseases classified elsewhere
#> 25                                                      I51.6 Cardiovascular disease, unspecified
#> 26                                                                             I51.7 Cardiomegaly
#> 27                                                                  I97.0 Postcardiotomy syndrome
#> 28                                                                  I98.0 Cardiovascular syphilis
#> 29 I98.1 Cardiovascular disorders in other infectious and parasitic diseases classified elsewhere
#> 30                                                         O90.3 Cardiomyopathy in the puerperium
#> 31                               P29 Cardiovascular disorders originating in the perinatal period
#> 32                       P29.8 Other cardiovascular disorders originating in the perinatal period
#> 33                 P29.9 Cardiovascular disorder originating in the perinatal period, unspecified
#> 34                                                                        R57.0 Cardiogenic shock
#> 35                                      R94.3 Abnormal results of cardiovascular function studies
#> 36                          T46 Poisoning by agents primarily affecting the cardiovascular system
#> 37               T46.9 Other and unspecified agents primarily affecting the cardiovascular system
#> 38                                       Y52 Agents primarily affecting the cardiovascular system
#> 39               Y52.9 Other and unspecified agents primarily affecting the cardiovascular system
#> 40                                   Y71 Cardiovascular devices associated with adverse incidents
#> 41                                  Z03.5 Observation for other suspected cardiovascular diseases
#> 42                               Z13.6 Special screening examination for cardiovascular disorders